navbar content moved to content
