+++ src/content/nav.js
@@ -1,0 +1,39 @@
+export const content = {
+  pages: [
+    {
+      name: 'Home',
+      link: '/',
+      description: 'Welcome to BUILD UMass!',
+    },
+    {
+      name: 'About',
+      longName: 'About Us',
+      link: '/about',
+      description: 'Learn about our mission, story, and team.',
+    },
+    {
+      name: 'Services',
+      longName: 'Our Services',
+      link: '/services',
+      description: 'The services we provide to our community.',
+    },
+    {
+      name: 'Projects',
+      longName: 'Our Projects',
+      link: '/projects',
+      description: 'Take a look through our past and present projects.',
+    },
+    {
+      name: 'Apply',
+      longName: 'Join Us',
+      link: '/apply',
+      description: 'Join our team and help us make a difference.',
+    },
+    {
+      name: 'Contact',
+      longName: 'Contact Us',
+      link: 'https://forms.gle/uuA2s98v4oC4o8TU7',
+      description: 'Get in touch with us.',
+    },
+  ]
+}
