+++ src/pages/_app.js
@@ -2,7 +2,7 @@
 import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
 import '../styles/style.scss';
 import Footer from '../components/Footer';
-import Navbar from '../components/Navbar';
+import Navbar from '../components/nav/Navbar';
 
 // Globalize CSS
 export default function MyApp({ Component, pageProps }) {
