+++ src/pages/about.js
@@ -5,7 +5,6 @@ import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Button from 'react-bootstrap/Button';
-import Navbar from '../components/Navbar';
 import JumbotronHeader from '../components/JumbotronHeader.jsx';
 
 const about = () => {
