+++ src/styles/style.scss
@@ -578,7 +578,8 @@ a {
 }
 
 .pg-jumbotron {
-	min-height: 50vh;
+	height: 70vh;
+  max-height: 550px;
 }
 
 .diagonal {
